+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 
 Todas las modificaciones y mejoras del sitio serán registradas aquí.
 
-## [1.0.0] - 2025-08-28
-- Versión inicial registrada en el changelog.
+
+## [28-08-2025]
+- Se muestran solo tres testimonios breves y funcionales de músicos en la sección principal.
+- Se agregó botón "Saber más" para ver testimonios.
+- Se agregaron páginas de Políticas y Privacidad, y Términos y Condiciones.
+- Se eliminó el link duplicado de "Ver todos los modelos" en Instrumentos Destacados.